Moved more main menu code into UI model and controller namespaces.
